feat!: [#878] extract net section in core config section
