+++ src/lib.rs
@@ -172,12 +172,10 @@
 //!
 //! [core]
 //! announce_interval = 120
-//! external_ip = "0.0.0.0"
 //! inactive_peer_cleanup_interval = 600
 //! max_peer_timeout = 900
 //! min_announce_interval = 120
 //! mode = "public"
-//! on_reverse_proxy = false
 //! persistent_torrent_completed_stat = false
 //! remove_peerless_torrents = true
 //! tracker_usage_statistics = true
@@ -185,6 +183,10 @@
 //!   [core.database]
 //!   driver = "Sqlite3"
 //!   path = "./storage/tracker/lib/database/sqlite3.db"
+//!
+//!   [core.net]
+//!   external_ip = "0.0.0.0"
+//!   on_reverse_proxy = false
 //!
 //! [[udp_trackers]]
 //! bind_address = "0.0.0.0:6969"
